+++ app/build.gradle
@@ -123,7 +123,6 @@ dependencies {
     implementation libs.commons
 
     implementation "com.google.accompanist:accompanist-systemuicontroller:0.18.0"
-    implementation 'com.google.firebase:firebase-storage:20.0.0'
     implementation "androidx.datastore:datastore-preferences:1.0.0"
 }
 kapt {
