+++ app/src/main/java/com/example/comusenias/domain/models/state/ChangeProfileState.kt
@@ -3,6 +3,11 @@ package com.example.comusenias.domain.models.state
 import com.example.comusenias.presentation.ui.theme.EMPTY_STRING
 
 data class ChangeProfileState(
-    var userName: String = EMPTY_STRING,
-    var image: String? = EMPTY_STRING
+    var id: String = EMPTY_STRING,
+    var name: String = EMPTY_STRING,
+    var tel: String = EMPTY_STRING,
+    var email: String = EMPTY_STRING,
+    var image: String? = EMPTY_STRING,
+    var date: String = EMPTY_STRING,
+    var idSpecialist: String = EMPTY_STRING,
 )
