+++ app/src/main/java/com/example/comusenias/presentation/component/profile/ChildrenProfileContent.kt
@@ -1,6 +1,6 @@
 package com.example.comusenias.presentation.component.profile
 
-
+import android.Manifest
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -13,7 +13,9 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.DateRange
 import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material.icons.filled.Email
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -28,40 +30,36 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavHostController
 import coil.compose.AsyncImage
 import com.example.comusenias.R
 import com.example.comusenias.presentation.component.defaults.DialogCapturePicture
 import com.example.comusenias.presentation.component.defaults.app.TextFieldApp
-import com.example.comusenias.presentation.component.defaults.app.TextFieldAppPassword
-import com.example.comusenias.presentation.ui.theme.CHANGE_PROFILE_CONFIRM_PASSWORD
-import com.example.comusenias.presentation.ui.theme.CHANGE_PROFILE_PASSWORD
-import com.example.comusenias.presentation.ui.theme.CHANGE_PROFILE_USER
 import com.example.comusenias.presentation.ui.theme.IMAGE_AVATAR
 import com.example.comusenias.presentation.ui.theme.IMAGE_EDIT_BUTTON
+import com.example.comusenias.presentation.ui.theme.PROFILE_BIRTHDATE
+import com.example.comusenias.presentation.ui.theme.PROFILE_EMAIL
+import com.example.comusenias.presentation.ui.theme.PROFILE_NAME_SURNAME
 import com.example.comusenias.presentation.ui.theme.SELECTED_IMAGE
 import com.example.comusenias.presentation.ui.theme.SIZE140
 import com.example.comusenias.presentation.ui.theme.SIZE55
 import com.example.comusenias.presentation.ui.theme.size10
 import com.example.comusenias.presentation.ui.theme.size20
-import com.example.comusenias.presentation.view_model.ChangeProfileViewModel
+import com.example.comusenias.presentation.view_model.ChildrenProfileViewModel
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
 
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
-fun ChangeProfileContent(
-    navController: NavHostController,
-    modifier: Modifier,
-    viewModel: ChangeProfileViewModel = hiltViewModel()
+fun ChildrenProfileContent(
+    viewModel: ChildrenProfileViewModel = hiltViewModel()
 ) {
+
     // Camera permission state
     val cameraPermissionState = rememberPermissionState(
-        android.Manifest.permission.CAMERA
+        Manifest.permission.CAMERA
     )
 
-    val state = viewModel.state
     val dialogState = remember {
         mutableStateOf(false)
     }
@@ -75,8 +73,8 @@ fun ChangeProfileContent(
             viewModel.pickImage()
         }
     )
-
     viewModel.resultingActivityHandler.handle()
+
     Box(modifier = Modifier.padding(size20.dp)) {
 
         Column(
@@ -102,11 +100,11 @@ fun ChangeProfileContent(
                         },
                 ) {
 
-                    if (viewModel.state.image?.isNotEmpty() == true) {
+                    if (viewModel.userData.image?.isNotEmpty() == true) {
                         AsyncImage(
                             modifier = Modifier
                                 .size(SIZE140.dp), contentScale = ContentScale.Crop,
-                            model = viewModel.state.image, contentDescription = SELECTED_IMAGE
+                            model = viewModel.userData.image, contentDescription = SELECTED_IMAGE
                         )
                     } else {
                         Image(
@@ -133,34 +131,42 @@ fun ChangeProfileContent(
 
             Spacer(modifier = Modifier.height(SIZE55.dp))
 
+            Spacer(modifier = Modifier.height(size20.dp))
             TextFieldApp(
-                value = state.name,
-                onValueChange = { viewModel.onUsernameInput(it) },
-                validateField = { viewModel.validateUserName() },
-                label = CHANGE_PROFILE_USER,
+                icon = Icons.Default.Email,
+                value = viewModel.userData.email,
+                onValueChange = {},
+                validateField = {},
+                label = PROFILE_EMAIL,
                 keyboardType = KeyboardType.Text,
-                errorMsg = viewModel.errorUserName,
-                icon = Icons.Default.Person
+                hideText = false,
+                readOnly = true
             )
-
-            Spacer(modifier = Modifier.height(size20.dp))
-
-            TextFieldAppPassword(
-                label = CHANGE_PROFILE_PASSWORD,
-                value = CHANGE_PROFILE_PASSWORD,
+            Spacer(modifier = Modifier.height(10.dp))
+            TextFieldApp(
+                icon = Icons.Default.Person,
+                value = viewModel.userData.name,
                 onValueChange = {},
                 validateField = {},
+                label = PROFILE_NAME_SURNAME,
+                keyboardType = KeyboardType.Text,
+                hideText = false,
+                readOnly = true
             )
-
-            Spacer(modifier = Modifier.height(size20.dp))
-            TextFieldAppPassword(
-                label = CHANGE_PROFILE_CONFIRM_PASSWORD,
-                value = CHANGE_PROFILE_PASSWORD,
+            Spacer(modifier = Modifier.height(10.dp))
+            TextFieldApp(
+                icon = Icons.Default.DateRange,
+                value = viewModel.userData.date,
                 onValueChange = {},
                 validateField = {},
-
-                )
-            Spacer(modifier = Modifier.height(size10.dp))
+                label = PROFILE_BIRTHDATE,
+                keyboardType = KeyboardType.Text,
+                hideText = false,
+                readOnly = true
+            )
         }
+
     }
+    SaveImageChildrenProfile()
+
 }
